Split out status
